intial up load of classes and some data.
This is the first upload of the new repo, We had a mishap of things not wanting to sync with wills and Joes, stuff. so as per nevin it would of been easier to just create a new repo and have one person upload what we have and then go from there.