FInal Revamp after getting database to work
I was having problems getting certain things in my Queries to work with the data base after our SNAFU yesterday with the database. I got the Data base working, by completely deleting the data base and reattaching a new database to the VS code. Then anywhere the data base was referenced i had to change it to HumaneSociety2 because that was the name of the new data base.